+++ Sources/KeyboardKit/Callouts/InputCalloutStyle.swift
@@ -39,16 +39,8 @@ public struct InputCalloutStyle {
 public extension InputCalloutStyle {
     
     /**
-     This is the standard style that will be used by default.
+     This is the standard callout style that will be used by
+     default. It looks like a system keyboard callout.
      */
     static var standard = InputCalloutStyle()
-    
-    /**
-     This is the standard style for a system styled callout.
-     */
-    static func systemStyle(for context: KeyboardContext) -> InputCalloutStyle {
-        var style = InputCalloutStyle.standard
-        style.callout = .systemStyle(for: context)
-        return style
-    }
 }