+++ RELEASE_NOTES.md
@@ -19,6 +19,7 @@ The dark appearance colors have been renamed. The old names are now deprecated.
 ### ✨ New features
 
 * `AudioFeedback` and `AudioFeedbackConfiguration` is now `Codable`.
+* `HapticFeedback` and `HapticFeedbackConfiguration` is now `Codable`.
 * `KeyboardLocale` is now `Codable`.
 * `KeyboardColor` has new colors.
 * `Color+Keyboard` has new colors.